+++ lib/src/renderer/liquid_glass_renderer.dart
@@ -14,7 +14,11 @@ export '../engine/liquid_glass.dart' show LiquidGlass;
 export '../engine/liquid_glass_blend_group.dart' show LiquidGlassBlendGroup;
 export '../engine/liquid_glass_layer.dart' show LiquidGlassLayer;
 export '../engine/liquid_glass_settings.dart'
-    show GlassBodyMode, LiquidGlassSettings, PlatformViewGlassMode;
+    show
+        GlassBodyMode,
+        GlassLensModel,
+        LiquidGlassSettings,
+        PlatformViewGlassMode;
 export '../engine/liquid_shape.dart';
 export '../engine/stretch.dart'
     show